Validate that a field is not equal to a specified value

The sv_not_equal() function compares the field value to a specified value with the != operator.

sv_not_equal(
  rhs,
  message_fmt = "Must not be equal to {rhs}.",
  allow_multiple = FALSE,
  allow_na = FALSE,
  allow_nan = FALSE,
  allow_inf = FALSE
)

Arguments

rhs
The right hand side (RHS) value is to be used for the comparison with the field value. The validation check will effectively be of the form <field> != <rhs>.

message_fmt
The validation error message to use if the field fails the validation test. Use the "{rhs}" string parameter to customize the message, including what was set in rhs. While the default message uses this string parameter, it is not required in a user-defined message_fmt string.

allow_multiple
If FALSE (the default), then the length of the input vector must be exactly one; if TRUE, then any length is allowed (including a length of zero; use sv_required() if one or more values should be required).

allow_na
If FALSE (the default for both options), then any NA or NaN element will cause validation to fail.

allow_nan
If FALSE (the default for both options), then any NA or NaN element will cause validation to fail.

allow_inf
If FALSE (the default), then any Inf or -Inf element will cause validation to fail.

Value

A function suitable for use as an InputValidator$add_rule() rule.

See also

Examples

## Only run examples in interactive R sessions
if (interactive()) {

library(shiny)
library(shinyvalidate)

ui <- fluidPage(
  textInput("score", "Number")
)

server <- function(input, output, session) {
  
  # Validation rules are set in the server, start by
  # making a new instance of an `InputValidator()`
  iv <- InputValidator$new()

  # Basic usage: `sv_not_equal()` requires a value
  # to compare against the field value; a message
  # will be shown if the validation of
  # `input$score` fails
  iv$add_rule("score", sv_not_equal(0))

  # Finally, `enable()` the validation rules
  iv$enable()
}

shinyApp(ui, server)

}